pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reader"
version = "1.0.0"
description = "Lab instrumentation data pipeline (plugin engine)"
readme = "README.md"
requires-python = ">=3.12,<3.13"
authors = [{ name = "Eric J. South", email = "ericjohnsouth@gmail.com" }]


dependencies = [
  "dnadesign",
  "pandas>=2.1",
  "pydantic>=2.6",
  "pyyaml>=6.0.2",
  "typer>=0.12.0",
  "rich>=13.7",
  "pyarrow>=15.0",
  "openpyxl>=3.1",
  "seaborn>=0.13",
  "pillow>=10.0",
  "flowio>=1.4",
]

[project.scripts]
reader = "reader.core.cli:app"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
include = ["reader*"]
namespaces = false

[dependency-groups]
dev = [
  "pytest>=8.2",
  "ruff>=0.4",
  "pre-commit>=3.7",
]

notebooks = [
  "marimo>=0.18.4",
]

[tool.uv]
required-version = ">=0.9.18,<0.10"

[tool.uv.sources]
dnadesign = { git = "https://github.com/e-south/dnadesign", branch = "main" }

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src"]
extend-exclude = [".venv", "build", "dist", ".eggs", "*.egg-info"]

[tool.ruff.lint]
select = ["E","F","W","I","UP","B","SIM","C4","PLC","PLE"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101","PLR2004","B011"]
"**/*.ipynb" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false
docstring-code-format = true

[tool.pytest.ini_options]
testpaths = [
  "src/reader/tests",
  "tests"
]

python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*", "*Tests"]
python_functions = ["test_*"]

addopts = [
  "--strict-config",
  "--strict-markers",
  "-ra"
]

norecursedirs = [
  ".venv", "build", "dist", ".eggs", "*.egg-info",
  "outputs", "plots", "raw_data", "inputs", "_scratch"
]
